Added delete all data button and modal
