+++ include/raylibhelper.h
@@ -8,6 +8,8 @@
 
 
 void draw_about(void);
+void create_trainer_name_str(const struct TrainerInfo *trainer, char *trainer_name, bool show_gender);
+void create_trainer_id_str(const struct TrainerInfo *trainer, char *trainer_id);
 void draw_change_dir(struct SaveFileData *save_file_data);
 void draw_settings(void);
 void draw_main_menu(struct SaveFileData *save_file_data);
